Ignoring slow funciton during startup
During startup the node will synchronize with the blockchain, while
doing so it will fecth large batches of events that have to be decoded
and converted to internal state changes, which will take considerable
time. To avoid crashing the node during initialization the monitoring is
only enable after the runnables have started.
